created hooks to Combobox component
