+++ src/app/components/ComboboxCountries/useNextAndPrevDay.ts
@@ -1,0 +1,32 @@
+import { DateArray } from '@/types/DateArray.types'
+import { useEffect, useState } from 'react'
+
+export const useNextAndPrevDay = (
+  filteredDates: DateArray[],
+  currentDate: Date,
+) => {
+  const [isNextDate, setIsNextDate] = useState<boolean>(false)
+  const [isPrevDate, setIsPrevDate] = useState<boolean>(false)
+  useEffect(() => {
+    const thereIsNextOrPrevDate = (
+      datesArray: DateArray[],
+      currentDate: Date,
+    ) => {
+      const validationNextDay = datesArray.some(([, groupedDate]) => {
+        const date = new Date(groupedDate.date)
+        const isNextDate = date.getDate() === currentDate.getDate() + 1
+        return isNextDate
+      })
+      const validationPrevDay = datesArray.some(([, groupedDate]) => {
+        const date = new Date(groupedDate.date)
+        const isPrevDate = date.getDate() === currentDate.getDate() - 1
+        return isPrevDate
+      })
+      setIsNextDate(validationNextDay)
+      setIsPrevDate(validationPrevDay)
+    }
+    thereIsNextOrPrevDate(filteredDates, currentDate)
+  }, [filteredDates, currentDate])
+
+  return { isNextDate, isPrevDate }
+}
