+++ src/types/DateArray.types.ts
@@ -1,0 +1,8 @@
+export type DateArray = [
+  string,
+  {
+    countryCodes: [string[]]
+    gmt?: string
+    date: string
+  },
+]
